Modularize JavaScript generator
