+++ .gitignore
@@ -49,3 +49,4 @@ docs/_build/
 *.backup
 *~
 *.swp
+backups/
