+++ generators/__init__.py
@@ -1,5 +1,7 @@
-"""Generators package"""
+"""Generators package."""
+
+from .css_lira import LiraCSSGenerator
 from .html_lira import LiraHTMLGenerator
 from .index_gen import IndexGenerator
 
-__all__ = ['LiraHTMLGenerator', 'IndexGenerator']
+__all__ = ["LiraCSSGenerator", "LiraHTMLGenerator", "IndexGenerator"]
