+++ generators/js/__init__.py
@@ -1,0 +1,6 @@
+"""JavaScript generation helpers."""
+
+from .generator import JavaScriptGenerator
+from .serializer import PhaseSerializer
+
+__all__ = ["JavaScriptGenerator", "PhaseSerializer"]
