+++ scripts/test_log_20250920_141006.txt
@@ -1,0 +1,67 @@
+============================================================
+  KING LEAR COMIC GENERATOR - 12 CHARACTERS
+============================================================
+[OK] Copied training.html to output
+[FOUND] king_lear.json ✓
+[FOUND] cordelia.json ✓
+[FOUND] goneril.json ✓
+[FOUND] regan.json ✓
+[FOUND] gloucester.json ✓
+[FOUND] edgar.json ✓
+[FOUND] edmund.json ✓
+[FOUND] kent.json ✓
+[FOUND] fool.json ✓
+[FOUND] albany.json ✓
+[FOUND] cornwall.json ✓
+[FOUND] oswald.json ✓
+
+[INFO] Processing 12 characters...
+
+[GENERATING] king_lear...
+[OK] Saved: king_lear.html
+
+[GENERATING] cordelia...
+[OK] Saved: cordelia.html
+
+[GENERATING] goneril...
+[OK] Saved: goneril.html
+
+[GENERATING] regan...
+[OK] Saved: regan.html
+
+[GENERATING] gloucester...
+[OK] Saved: gloucester.html
+
+[GENERATING] edgar...
+[OK] Saved: edgar.html
+
+[GENERATING] edmund...
+[OK] Saved: edmund.html
+
+[GENERATING] kent...
+[OK] Saved: kent.html
+
+[GENERATING] fool...
+[OK] Saved: fool.html
+
+[GENERATING] albany...
+[OK] Saved: albany.html
+
+[GENERATING] cornwall...
+[OK] Saved: cornwall.html
+
+[GENERATING] oswald...
+[OK] Saved: oswald.html
+
+[GENERATING] Index page...
+[OK] Index page created
+
+============================================================
+[SUMMARY]
+  • Characters found: 12/12
+  • Pages generated: 12
+  • Output location: /workspace/comic-website/output
+  • Status: ✅ ALL CHARACTERS READY!
+============================================================
+
+[OK] Opening in browser...
