+++ utils/__init__.py
@@ -1,0 +1,30 @@
+"""Utility helpers for the King Lear Comic generator."""
+
+from .console_output import console_line, info, warning, error, success
+from .file_operations import (
+    read_text,
+    write_text,
+    read_json,
+    write_json,
+    ensure_directory,
+)
+from .text_processing import collapse_whitespace, slugify, split_sentences
+from .validation import ensure_file_length, ensure_directory_structure
+
+__all__ = [
+    "console_line",
+    "info",
+    "warning",
+    "error",
+    "success",
+    "read_text",
+    "write_text",
+    "read_json",
+    "write_json",
+    "ensure_directory",
+    "collapse_whitespace",
+    "slugify",
+    "split_sentences",
+    "ensure_file_length",
+    "ensure_directory_structure",
+]
